+++ src/imio/helpers/content.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from imio.helpers import HAS_PLONE_5_AND_MORE
+from imio.helpers import HAS_PLONE_6_AND_MORE
 from imio.helpers.interfaces import IContainerOfUnindexedElementsMarker
 from imio.helpers.workflow import do_transitions
 from imio.pyutils.utils import safe_encode  # noqa, temporary import for backward compatibility as was also defined here
@@ -31,7 +31,7 @@ import logging
 import os
 
 
-if HAS_PLONE_5_AND_MORE:
+if HAS_PLONE_6_AND_MORE:
     from Products.CMFPlone.interfaces import IEditingSchema
 
 logger = logging.getLogger('imio.helpers.content')