+++ src/imio/helpers/utils.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from imio.helpers import HAS_PLONE_5_AND_MORE
+from imio.helpers import HAS_PLONE_6_AND_MORE
 from plone import api
 from plone.namedfile.file import NamedBlobImage
 
@@ -8,7 +8,7 @@ from plone.namedfile.file import NamedBlobImage
 def create_image_content(
     container, title, id, data, filename=None, portal_type="Image"
 ):
-    if HAS_PLONE_5_AND_MORE:
+    if HAS_PLONE_6_AND_MORE:
         obj = api.content.create(
             container=container,
             type=portal_type,